event usage logic is a link everything together pain
