+++ frontend/src/scenes/session-recordings/player/inspector/playerInspectorLogic.ts
@@ -7,7 +7,6 @@ import { TaxonomicFilterGroupType } from 'lib/components/TaxonomicFilter/types'
 import { Dayjs, dayjs } from 'lib/dayjs'
 import { featureFlagLogic } from 'lib/logic/featureFlagLogic'
 import { eventToDescription, humanizeBytes, objectsEqual, toParams } from 'lib/utils'
-import { eventUsageLogic } from 'lib/utils/eventUsageLogic'
 import {
     InspectorListItemPerformance,
     performanceEventDataLogic,
@@ -21,6 +20,7 @@ import {
     convertUniversalFiltersToRecordingsQuery,
     MatchingEventsMatchType,
 } from 'scenes/session-recordings/playlist/sessionRecordingsPlaylistLogic'
+import { sessionRecordingEventUsageLogic } from 'scenes/session-recordings/sessionRecordingEventUsageLogic'
 
 import { RecordingsQuery } from '~/queries/schema/schema-general'
 import { getCoreFilterDefinition } from '~/taxonomy/helpers'
@@ -220,7 +220,7 @@ export const playerInspectorLogic = kea<playerInspectorLogicType>([
         actions: [
             miniFiltersLogic,
             ['setMiniFilter', 'setSearchQuery'],
-            eventUsageLogic,
+            sessionRecordingEventUsageLogic,
             ['reportRecordingInspectorItemExpanded'],
             sessionRecordingDataLogic(props),
             ['loadFullEventData', 'setTrackedWindow'],
@@ -1015,7 +1015,7 @@ export const playerInspectorLogic = kea<playerInspectorLogicType>([
         setItemExpanded: ({ index, expanded }) => {
             if (expanded) {
                 const item = values.items[index]
-                eventUsageLogic.actions.reportRecordingInspectorItemExpanded(item.type, index)
+                actions.reportRecordingInspectorItemExpanded(item.type, index)
 
                 if (item.type === FilterableInspectorListItemTypes.EVENTS) {
                     actions.loadFullEventData(item.data)
